Update and rename Sudoku.py to solver.py
A multi year journey comes to an end. Maybe my instinctual method to solve a Sudoku puzzle was novel, but not the most efficient. Bit-masking was a way easier method of tracking used numbers then my list look up. It was a challenge for me to even implement all the different techniques that made use of Cell.possible_values. That method also didn't guarantee a correct solution on the hardest puzzle, I knew for a fact those needed guessing and undoing, and I struggled coming up with that algorithm myself.

But now I've come away from this with the knowledge of bit masking, encoding flags and states into integers. Very cool thing I want to play with more.